server/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
python-dotenv==1.0.1
PyMySQL==1.1.1
SQLAlchemy==2.0.36
typing_extensions==4.12.2
greenlet==3.0.3
